Refactor streams use in tarball fetcher (#348)
* Use built-in `stream.Transform` to create `HashStream`s

Since we're only targeting newer Node versions with a reasonable stream module, and we can use the `class` syntax sugar, it makes sense to extend `stream.Transform` directly to create a `HashStream`. In a future commit, this will allows to remove `through2` altogether.

Also, `HashStream` was previously mistyped as a `stream$Readable`, when in fact, it's a `stream$Transform`.

* Fix `cachedStream` type

`validateStream` (`cachedStream` is piped into it a few lines below) was mistyped as a `stream$Readable` instead of a `stream$Transform`. It's now correctly typed, so this isn't necessary.

* Use idiomatic streams to unpack tarballs

Instead of doing weird nested stream passing, there's now an `Unpack` stream constructor that can unzip a stream if the input is gziped, or passthrough the data otherwise.

* Unnest tarball saving streams

* Use `stream.PassThrough` for logging instead of `through2`

* Remove `through2` dependency