Clarify mirror path flow types (#371)
It wasn't entirely clear when it was ok to have a "maybe" mirror vs. an "empty" string mirror. Fixes #350.

It's unfortunate that the prevalent js style for checking for nullness is to check for falseness. Because Flow lets you express "maybe" as `null` or `undefined`, using falsey checks instead of `!= null`, leads to a lot of loss of information.