Most Windows tests passing
There are two exceptions:

expect().toMatchSnapshot is CRLF-sensitive. Since that facility is
supplied by jest, I can't quickly fix it.

There is some quirkiness with copying relative symlinks on Windows. For
now, I guaraded the relevant test, but I'd like to figure out what
semantics we want and how to make it work.