Fix request authentication for scoped private registry tarballs (#1666)
* Fix request authentication for scoped private registry tarballs

Yarn correctly sends authorization headers when resolving metadata, but
when it's time to download the tarballs, it calls registry.request
with a full tarball url. This causes the getRegistry() function to
incorrectly return the DEFAULT_REGISTRY url which means the authorization
headers are no longer sent. The response in this case is a JSON error
object instead of a tarball file. Untaring fails with "invalid tar file"

* Don't assume anything about the registry URL, read registries from config

* Add a test for installing private packages requiring authentication

* Change the authed package testing strategy to be more local

* Include non scoped registries in available registries list

* Unnest the authed pkg test, use FlowFixMe for afterEach