Add usage section to README.me with link to docs (#1950)
Upon looking for basic usage commands I discovered that there is no
mention of this in the official README.md, just an indirect link to
the installation guide section of the official documentation. I
therefore propose to add a separate usage section, and/or a
documentation section with url to the official documentation for
further reading.