bugfixes to yarn pack, retry of #1464 (#2062)
* add tests for pack with files-array

* include mandatory files if not in files-array

npm pack includes some files even if they are not included in the “files” field in package.json. This commit creates the same behaviour in yarn pack.

* explicitly exclude dotfiles in pack with files

if files is used in package.json all other files should be excluded. Files with dot as first character have to be excluded explicitly because minimatch will not match them by default.