Simplified test due to inability to mock process.platform (#2262)
The original test asserts that window cmd line formatting is properly applied or not applied, for only the target platform the tests are executed on - that will have to be good enough.

It is not possible to test both branches from one platform because the path module employed in the test alters its export based on process.platform in Node bootstrap, well before one can mock it.