Resolves issues with git+ssh when ssh daemon has banner set. (#2905)
* Silencing git+ssh banner

If your ssh daemon has a banner set this currently causes yarn add to fail because it’s expecting to get an array of git shas instead of the banner.  This uses the quiet option of ssh to exclude the banners for git+ssh.

* Fixing linting issues

Fixed linting issues found during circle ci build.