[feature] Ability to pass command CLI arguments via configuration in .yarnrc (#3033)
This PR opens the way to share configuration between multiple projects via the use of the .yarnrc
files. The current design it pretty simple: it simply adds the supported options from the yarnrc file at the beginning of the command line.

Some notes:

The yarnrc is currently parsed synchronously. I believe this is not an issue, since it's quite literally the very first thing we need to do when booting the application (even before the command line can be parsed), so there's really nothing we could run concurrently anyway.

The yarnrc is read in another file (https://github.com/yarnpkg/yarn/blob/master/src/registries/yarn-registry.js). I haven't changed how it works there, since I wanted my first draft to have as few changes as possible, but it's something that should probably be fixed so that the code only use one path to manage the yarnrc files.